+++ src/main/java/com/alkemy/ong/security/config/AppSecurity.java
@@ -3,9 +3,7 @@ package com.alkemy.ong.security.config;
 import com.alkemy.ong.exception.CustomAccessDeniedHandler;
 import com.alkemy.ong.exception.CustomAuthenticationEntryPoint;
 import com.alkemy.ong.security.filter.JwtRequestFilter;
-import com.alkemy.ong.security.service.impl.UserServiceImpl;
-
-import static com.alkemy.ong.util.Constants.ROLE_USER;
+import com.alkemy.ong.security.service.impl.UserDetailsCustomServiceImpl;
 import static com.alkemy.ong.util.Constants.ALL_ROLES;
 import static com.alkemy.ong.util.Constants.Endpoints.*;
 import static com.alkemy.ong.util.Constants.ROLE_ADMIN;
@@ -27,14 +25,14 @@ import org.springframework.security.web.authentication.UsernamePasswordAuthentic
 
 import javax.servlet.Filter;
 import org.springframework.http.HttpMethod;
-
+import org.springframework.security.core.userdetails.UserDetailsService;
 
 @Configuration
 @EnableWebSecurity
 public class AppSecurity extends WebSecurityConfigurerAdapter {
 
     @Autowired
-    private UserServiceImpl userDetailsCustomService;
+    private UserDetailsService userDetailsService;
 
     @Autowired
     private PasswordEncoder encoder;
@@ -48,18 +46,18 @@ public class AppSecurity extends WebSecurityConfigurerAdapter {
                 .authorizeRequests()
                 .antMatchers(API_UI_ANTMATCHER, API_DESCRIPTION_ANTMATCHER).permitAll()
                 .and().authorizeRequests()
-                .antMatchers("/v2/api-docs", "/v3/api-docs","/swagger-ui.html", "api/docs").permitAll()
-                .antMatchers("/auth/*").permitAll()
+                .antMatchers(DOCS).permitAll()
+                .antMatchers(AUTH_ALL).permitAll()
                 .antMatchers(HttpMethod.GET, USER).hasAnyAuthority(ROLE_ADMIN)
                 .antMatchers(HttpMethod.PATCH, USER_ID).hasAnyAuthority(ALL_ROLES)
                 .antMatchers(HttpMethod.DELETE, USER_ID).hasAnyAuthority(ALL_ROLES)
-                .antMatchers(HttpMethod.PUT, "/news/{id}").hasAnyAuthority(ROLE_ADMIN)
-                .antMatchers(HttpMethod.DELETE, "/news/{id}").hasAnyAuthority(ROLE_ADMIN)
-                .antMatchers(HttpMethod.GET, "/news/{id}").hasAnyAuthority(ROLE_ADMIN)
-                .antMatchers(HttpMethod.GET, "/news**").permitAll()
-                .antMatchers(HttpMethod.DELETE, "/categories/{id}").hasAnyAuthority(ROLE_ADMIN)
-                .antMatchers(HttpMethod.PUT, "/categories/{id}").hasAnyAuthority(ROLE_ADMIN)
-                .antMatchers(HttpMethod.GET, "/categories/{id}").hasAnyAuthority(ROLE_ADMIN)
+                .antMatchers(HttpMethod.PUT, NEWS_ID).hasAnyAuthority(ROLE_ADMIN)
+                .antMatchers(HttpMethod.DELETE, NEWS_ID).hasAnyAuthority(ROLE_ADMIN)
+                .antMatchers(HttpMethod.GET, NEWS_ID).hasAnyAuthority(ROLE_ADMIN)
+                .antMatchers(HttpMethod.GET, NEWS_ALL).permitAll()
+                .antMatchers(HttpMethod.DELETE, CATEGORY_ID).hasAnyAuthority(ROLE_ADMIN)
+                .antMatchers(HttpMethod.PUT, CATEGORY_ID).hasAnyAuthority(ROLE_ADMIN)
+                .antMatchers(HttpMethod.GET, CATEGORY_ID).hasAnyAuthority(ROLE_ADMIN)
                 .antMatchers(HttpMethod.GET, SLIDE).hasAuthority(ROLE_ADMIN)
                 .antMatchers(HttpMethod.GET, SLIDE_ID).hasAuthority(ROLE_ADMIN)
                 .antMatchers(HttpMethod.PUT, SLIDE_ID).hasAuthority(ROLE_ADMIN)
@@ -69,7 +67,7 @@ public class AppSecurity extends WebSecurityConfigurerAdapter {
                 .antMatchers(HttpMethod.PATCH, "/organization/public/{id}").hasAnyAuthority(ROLE_ADMIN)
                 .antMatchers(HttpMethod.PATCH, "/organization/{id}").hasAnyAuthority(ROLE_ADMIN)
                 .antMatchers(HttpMethod.POST, "/activities").hasAnyAuthority(ROLE_ADMIN)
-                .antMatchers(HttpMethod.POST, "/categories").hasAnyAuthority(ROLE_ADMIN)
+                .antMatchers(HttpMethod.POST, CATEGORY).hasAnyAuthority(ROLE_ADMIN)
                 .antMatchers(HttpMethod.DELETE, "/testimonials/{id}").hasAnyAuthority(ROLE_ADMIN)
                 .antMatchers(HttpMethod.GET, MEMBER).hasAnyAuthority(ROLE_ADMIN)
                 .antMatchers(HttpMethod.GET, "/comments").hasAnyAuthority(ROLE_ADMIN)
@@ -82,13 +80,12 @@ public class AppSecurity extends WebSecurityConfigurerAdapter {
                 .and().sessionManagement()
                 .sessionCreationPolicy(SessionCreationPolicy.STATELESS);
 
-
         httpSecurity.addFilterBefore((Filter) jwtRequestFilter, UsernamePasswordAuthenticationFilter.class);
     }
 
     @Override
     protected void configure(AuthenticationManagerBuilder auth) throws Exception {
-        auth.userDetailsService(userDetailsCustomService).passwordEncoder(passwordEncoder());
+        auth.userDetailsService(new UserDetailsCustomServiceImpl()).passwordEncoder(passwordEncoder());
     }
 
     @Bean
@@ -110,5 +107,10 @@ public class AppSecurity extends WebSecurityConfigurerAdapter {
     @Bean
     public AuthenticationEntryPoint authenticationEntryPoint() {
         return new CustomAuthenticationEntryPoint();
+    }
+    
+    @Bean
+    public UserDetailsService userDetailsService(){
+        return new UserDetailsCustomServiceImpl();
     }
 }
