+++ src/main/java/com/alkemy/ong/security/controller/AuthController.java
@@ -1,10 +1,11 @@
 package com.alkemy.ong.security.controller;
 
 import com.alkemy.ong.security.dto.*;
-import com.alkemy.ong.security.service.IUserService;
 import org.springframework.http.HttpStatus;
 import com.alkemy.ong.security.dto.AuthRequest;
 import com.alkemy.ong.security.dto.AuthResponse;
+import com.alkemy.ong.security.service.IAuthService;
+import com.alkemy.ong.util.Constants.Endpoints;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
@@ -14,21 +15,21 @@ import javax.validation.Valid;
 import java.io.IOException;
 
 @RestController
-@RequestMapping("/auth")
+@RequestMapping(Endpoints.AUTH)
 public class AuthController {
 
     @Autowired
-    private IUserService userService;
+    private IAuthService authService;
 
     @PostMapping("/register")
     public ResponseEntity<UserGetDto> register(@Valid @RequestBody UserPostDto dto) throws IOException {
-        UserGetDto userGetDto = userService.registerUser(dto);
+        UserGetDto userGetDto = authService.registerUser(dto);
         return ResponseEntity.status(HttpStatus.CREATED).body(userGetDto);
     }
 
     @PostMapping("/login")
     public ResponseEntity<AuthResponse> signIn(@Valid @RequestBody AuthRequest request) throws Exception {
-        AuthResponse response = userService.authenticate(request);
+        AuthResponse response = authService.authenticate(request);
 
         return ResponseEntity.ok(response);
 
@@ -36,7 +37,7 @@ public class AuthController {
 
     @GetMapping("/me")
     public ResponseEntity<UserInformationDto> userInformation(@Valid Authentication authentication) {
-        return ResponseEntity.ok(userService.getCurrentAuthenticatedUser(authentication));
+        return ResponseEntity.ok(authService.getCurrentAuthenticatedUser(authentication));
     }
 
 }
