+++ src/main/java/com/alkemy/ong/security/service/impl/UserDetailsCustomServiceImpl.java
@@ -1,0 +1,46 @@
+package com.alkemy.ong.security.service.impl;
+
+import com.alkemy.ong.security.model.Role;
+import com.alkemy.ong.security.model.User;
+import com.alkemy.ong.security.repository.UserRepository;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Locale;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.MessageSource;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+
+public class UserDetailsCustomServiceImpl implements UserDetailsService {
+    
+    @Autowired
+    private UserRepository userRepository;
+    
+    @Autowired
+    private MessageSource message;
+    
+    public UserDetailsCustomServiceImpl(){
+        
+    }
+    
+    @Override
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+
+        User user = userRepository.findByEmail(email);
+        if (user == null) {
+
+            throw new UsernameNotFoundException(message.getMessage("email.notFound", null, Locale.US));
+        }
+        List<GrantedAuthority> authorities = new ArrayList<>();
+
+        for (Role role : user.getRoles()) {
+            authorities.add(new SimpleGrantedAuthority(role.getName()));
+        }
+
+        org.springframework.security.core.userdetails.User userDetails = new org.springframework.security.core.userdetails.User(user.getEmail(), user.getPassword(), authorities);
+        return userDetails;
+    }
+}
